Modified setPoints and setTotalPoints for Assignment
Modified the bodies of setPoints and setTotalPoints for the Assignment class to make it more resilient against the type of value the Root returns.

Now it checks if the value the Root returns is really an Assignment, and to do nothing if it isn't. Otherwise, it does the modification as required.